Crash fixed when initializing with older dictionary representations
The older structure of the `dictionaryRepresentation` of an `SCRecordSession` represented the recordSegments as Strings, but the new structure is an `NSDictionary`. This was crashing our app when using the new version of `SCRecordSession` and `initWithDictionaryRepresentation`. I've added code to check and migrate the older structure to the new structure for apps that used older versions of SCRecorder.